==> notebooks/01_preparation_and_exploration.ipynb <==
<!-- **Phases 1-5: load, quality assesment, cleaning, feature engineering, visualization** -->
**Phases 1-3: load, quality assesment, cleaning**

## **Phase 1: Data Loading & Initial Exploration**

import pandas as pd

# Note: df is short for dataframe for pandas dataframe
df = pd.read_csv("../data/raw/data.csv")
df.head()

# Number of rows and columns
df.shape

# See column names and types
df.info()

# Conclusion:
# - id, Unnamed should be removed

## **Phase 2: Data Quality Assessment**

#  Count missing values in each column
df.isnull().sum()

# Check for duplicate rows
df.duplicated().sum()

## **Phase 3: Data Cleaning**

# Drop Unnamed column (useless) then check the info
df.drop(['Unnamed: 32'], axis=1, inplace=True)
df.info()

# Convert diagnosis from text to numbers
df['diagnosis'] = df['diagnosis'].map({'M': 1, 'B': 0})

# Check the conversion worked
print(df['diagnosis'].value_counts())
# Should show: 1 (malignant) and 0 (benign) counts

df.to_csv("../data/processed/cleaned_data.csv", index=False)
